Move files around to make it easy to embed sfold/ssync in other binaries
